+++ R/NMctl.R
@@ -3,7 +3,7 @@
 ##' @param x object to read.
 ##' @param ... Not used.
 ##' @return An object of class `NMctl`.
-##' @export
+##' @keywords internal
 readCtl <- function(x,...){
     UseMethod("readCtl",x)
 }
@@ -11,23 +11,28 @@ readCtl <- function(x,...){
 ##readCtl(x="run1.mod")
 
 ##' @import NMdata
-##' @export
+##' @keywords internal
 readCtl.character <- function(x,...){
-    ## ctl <- NMreadSection(file=x)
-    ctl <- readLines(x,warn=FALSE)
-    ## ctl <- as.NMctl(ctl)
-    class(ctl) <- "NMctl"
+    dots <- list(...)
+    if("lines"%in%names(dots)){
+        ### this only supports one ctl. Generalize?
+        x <- do.call(c,strsplit(x,"\\n"))
+        ctl <- x
+    } else {
+        ctl <- readLines(x,warn=FALSE)
+    }
+    ## class(ctl) <- "NMctl"
+    setattr(ctl,"class",unique(c("NMctl",class(ctl))))
     ctl
 }
 
-##' @export
+##' @keywords internal
 readCtl.NMctl <- function(x,...){
     x
 }
 
-##' @export
+##' @keywords internal
 print.NMctl <- function(x,...) {
-   
     if(!is.list(x)) x <- list(x)
     res <- lapply(x,function(x){
         x <- x[!grepl("^ *$",x)]
@@ -40,18 +45,17 @@ print.NMctl <- function(x,...) {
 ##' @param x object to convert
 ##' @param ... Not used
 ##' @return An object of class `NMctl`.
-##' @export
+##' @keywords internal
 as.NMctl <- function(x,...){
     UseMethod("as.NMctl",x)
 }
 
-##' @export
+##' @keywords internal
 as.NMctl.character <- function(x,...){
-    
-    readCtl(x)
+    readCtl(x,...)
 }
 
-##' @export
+##' @keywords internal
 as.NMctl.NMctl <- function(x,...){
     x
 }