+++ tests/testthat/test_simplePath.R
@@ -11,5 +11,8 @@ test_that("basic",{
     expect_equal(simplePath("c:/hel/wee/ ww"),"c:/hel/wee/ ww")
     expect_equal(simplePath("c:/hel/wee/  ww"),"c:/hel/wee/  ww")
 
+    expect_equal(simplePath("../hel/wee/ww"),"../hel/wee/ww")
+    expect_equal(simplePath("./hel/wee/ww"),"hel/wee/ww")
+
 })